requirements.txt
requests
beautifulsoup4
pandas
selenium
webdriver-manager
pytest
pymongo
nltk 
spacy 
networkx
aiohttp
python-bidi